Fix tests and make auto-discoverable
