+++ .editorconfig
@@ -1,0 +1,11 @@
+root = true
+
+[*]
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.{py,sh, bash}]
+indent_style = space
+indent_size = 4
+max_line_length = 99
