+++ ros2bag_tools/ros2bag_tools/test/conftest.py
@@ -1,4 +1,3 @@
-#!/usr/bin/python3
 # Copyright 2021 AIT Austrian Institute of Technology GmbH
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
