+++ rosbag2_tools/rosbag2_tools/test/test_flake8.py
@@ -21,7 +21,7 @@ import pytest
 @pytest.mark.flake8
 @pytest.mark.linter
 def test_flake8():
-    pkg_prefix = Path(__file__).parents[1]
+    pkg_prefix = Path(__file__).resolve().parents[2]
     config_file = pkg_prefix / '.flake8'
     rc, errors = main_with_errors(argv=['--config', str(config_file), str(pkg_prefix)])
     assert rc == 0, \
