+++ rosbag2_tools/setup.py
@@ -1,3 +1,4 @@
+from setuptools import find_packages
 from setuptools import setup
 
 package_name = 'rosbag2_tools'
@@ -5,7 +6,7 @@ package_name = 'rosbag2_tools'
 setup(
     name=package_name,
     version='0.1.0',
-    packages=[package_name],
+    packages=find_packages(exclude=['**.test']),
     data_files=[
         ('share/ament_index/resource_index/packages',
             ['resource/' + package_name]),
@@ -13,8 +14,10 @@ setup(
     ],
     install_requires=['setuptools'],
     zip_safe=True,
-    maintainer='Marcel Zeilinger',
-    maintainer_email='marcel.zeilinger@ait.ac.at',
+    author='Marcel Zeilinger',
+    author_email='marcel.zeilinger@ait.ac.at',
+    maintainer='Markus Hofstaetter',
+    maintainer_email='markus.hofstaetter@ait.ac.at',
     description='Python processing and visualization of ROS bags',
     license='Apache 2.0',
     tests_require=['pytest'],
